+++ packages/atomic/src/components/facets-v1/facet-value-box/facet-value-box.tsx
@@ -1,4 +1,5 @@
 import {FunctionalComponent, h} from '@stencil/core';
+import {createRipple} from '../../../utils/ripple';
 import {FacetValueProps} from '../facet-common';
 
 export const FacetValueBox: FunctionalComponent<FacetValueProps> = (
@@ -16,6 +17,7 @@ export const FacetValueBox: FunctionalComponent<FacetValueProps> = (
       <button
         part="value-box"
         onClick={() => props.onClick()}
+        onMouseDown={(e) => createRipple(e, {color: 'neutral'})}
         class={`value-box box-border w-full h-full flex flex-col items-center rounded text-on-background hover:border-primary-light focus:border-primary-light focus:outline-none p-2 group focus:bg-neutral-light hover:bg-neutral-light ${
           props.isSelected ? 'border-primary border-2' : 'border border-neutral'
         }`}